+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/UnusedPrivateMethodRule.java
@@ -21,6 +21,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodReference;
 import net.sourceforge.pmd.lang.java.ast.ASTModifierList;
 import net.sourceforge.pmd.lang.java.ast.ASTStringLiteral;
+import net.sourceforge.pmd.lang.java.ast.ASTThisExpression;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.ast.MethodUsage;
 import net.sourceforge.pmd.lang.java.ast.ModifierOwner.Visibility;
@@ -101,6 +102,25 @@ public class UnusedPrivateMethodRule extends AbstractIgnoredAnnotationRule {
                         && remainingUnused.remove(reffed) // note: side-effect
                         && remainingUnused.isEmpty()) {
                         consideredNames.remove(methodName); // clear this name
+                    }
+                }
+
+                // TODO workaround for missing method symbol resolution
+                // for some reason, we couldn't find the method declaration of the method call
+                // assume, it is in this class
+                // see test case: [java] UnusedPrivateMethod false-positive used in lambda #4817
+                if (reffed == null && ref instanceof ASTMethodCall) {
+                    ASTMethodCall call = (ASTMethodCall) ref;
+                    if (call.getQualifier() == null || call.getQualifier() instanceof ASTThisExpression) {
+                        Set<ASTMethodDeclaration> remainingUnused = consideredNames.get(methodName);
+
+                        // only remove this method, if we are sure, it's the correct and only overload
+                        if (remainingUnused.size() == 1) {
+                            ASTMethodDeclaration declaration = remainingUnused.iterator().next();
+                            if (declaration.getArity() == call.getArguments().size()) {
+                                consideredNames.remove(methodName);
+                            }
+                        }
                     }
                 }
             });